Add caching to the car service
This memory-layer caching is safe since the data should not be changed during the runtime of the app.
If the data was allowed to change I would first serve the cached copy and then fetch a fresh copy from the API and emit the new data using the observer (i.e. hot-swappable "stale-while-revalidate" approach).